+++ CHANGELOG.md
@@ -19,8 +19,8 @@ new minor version numbers. Please note that this is no version handling covered 
 
 ## 0.3.1 - Not released yet
 
-- Add support of unicode named classes `XID_Start` and `XID_Continue` by the help of the
-`unicode-xid` crate
+- Add support of lots of unicode named classes like `XID_Start` and `XID_Continue` by the help of
+the `seshat-unicode` crate
 
 ## 0.3.0 - 2024-08-29